Resource ID fragmentation related segfault fix.
git-svn-id: https://enigma-dev.svn.sourceforge.net/svnroot/enigma-dev/trunk@309 726d713c-bd1a-46cb-9ef0-1b6e075168bb

+++ ENIGMAsystem/SHELL/Universal_System/spritestruct.cpp
@@ -34,7 +34,7 @@ using namespace std;
 
 namespace enigma {
   sprite** spritestructarray;
-	extern unsigned int sprite_idmax;
+	extern size_t sprite_idmax;
 	sprite::sprite(): texturearray(NULL) {}
   sprite::sprite(unsigned int x): texturearray(new unsigned int[x]) {}
 }
